Inline applymap logic.
